// packages/wasm/lib/chili-wasm.d.ts
type EmbindString = ArrayBuffer|Uint8Array|Uint8ClampedArray|Int8Array|string;
export interface ClassHandle {
  isAliasOf(other: ClassHandle): boolean;
  delete(): void;
  deleteLater(): this;
  isDeleted(): boolean;
  // @ts-ignore - If targeting lower than ESNext, this symbol might not exist.
  [Symbol.dispose](): void;
  clone(): this;
}
export interface ShapeNode extends ClassHandle {
  get name(): string;
  set name(value: EmbindString);
  shape: TopoDS_Shape | undefined;
  get color(): string | undefined;
  set color(value: EmbindString | undefined);
  getChildren(): Array<ShapeNode>;
}

export interface Converter extends ClassHandle {
}

export interface ShapeResult extends ClassHandle {
  isOk: boolean;
  get error(): string;
  set error(value: EmbindString);
  shape: TopoDS_Shape;
}

export interface RemoveFilletResult extends ClassHandle {
  isOk: boolean;
  get error(): string;
  set error(value: EmbindString);
  shape: TopoDS_Shape;
  newEdges: Array<TopoDS_Shape>;
}

export interface ShapesResult extends ClassHandle {
  isOk: boolean;
  get error(): string;
  set error(value: EmbindString);
  shapes: Array<TopoDS_Shape>;
}
// ...
export interface TopAbs_ShapeEnumValue<T extends number> {
  value: T;
}
export type TopAbs_ShapeEnum = TopAbs_ShapeEnumValue<7>|TopAbs_ShapeEnumValue<6>|TopAbs_ShapeEnumValue<5>|TopAbs_ShapeEnumValue<4>|TopAbs_ShapeEnumValue<3>|TopAbs_ShapeEnumValue<2>|TopAbs_ShapeEnumValue<0>|TopAbs_ShapeEnumValue<1>|TopAbs_ShapeEnumValue<8>;

export interface TopAbs_OrientationValue<T extends number> {
  value: T;
}
export type TopAbs_Orientation = TopAbs_OrientationValue<0>|TopAbs_OrientationValue<1>|TopAbs_OrientationValue<2>|TopAbs_OrientationValue<3>;
// ...
export interface gp_Trsf extends ClassHandle {
  value(_0: number, _1: number): number;
  setValues(_0: number, _1: number, _2: number, _3: number, _4: number, _5: number, _6: number, _7: number, _8: number, _9: number, _10: number, _11: number): void;
}

export interface TopLoc_Location extends ClassHandle {
  transformation(): gp_Trsf;
  inverted(): TopLoc_Location;
}

export interface TopoDS extends ClassHandle {
}

export interface TopoDS_Shape extends ClassHandle {
  infinite(): boolean;
  isEqual(_0: TopoDS_Shape): boolean;
  isNull(): boolean;
  isPartner(_0: TopoDS_Shape): boolean;
  isSame(_0: TopoDS_Shape): boolean;
  getLocation(): TopLoc_Location;
  setLocation(_0: TopLoc_Location, _1: boolean): void;
  nbChildren(): number;
  nullify(): void;
  getOrientation(): TopAbs_Orientation;
  setOrientation(_0: TopAbs_Orientation): void;
  reverse(): void;
  reversed(): TopoDS_Shape;
  shapeType(): TopAbs_ShapeEnum;
  located(_0: TopLoc_Location, _1: boolean): TopoDS_Shape;
  move(_0: TopLoc_Location, _1: boolean): void;
  moved(_0: TopLoc_Location, _1: boolean): TopoDS_Shape;
}

export interface TColgp_Array1OfPnt extends ClassHandle {
  value(_0: number): gp_Pnt;
  setValue(_0: number, _1: gp_Pnt): void;
  length(): number;
}

export interface TopoDS_Vertex extends TopoDS_Shape {
}

export interface TopoDS_Edge extends TopoDS_Shape {
}

export interface TopoDS_Wire extends TopoDS_Shape {
}

export interface TopoDS_Face extends TopoDS_Shape {
}